% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix_fract_from_extended_segs.R
\name{admix_fract_from_extended_segs}
\alias{admix_fract_from_extended_segs}
\title{Calculate admixture fractions from the extended segments}
\usage{
admix_fract_from_extended_segs(E)
}
\arguments{
\item{E}{the extended segments tibble like what comes out of
\code{\link[=extend_ancestral_segments_3]{extend_ancestral_segments_3()}}. Basically a tibble with the columns
\code{indiv}, \code{chrom_f}, \code{copy_num}, \code{start}, \code{stop}.  The column \code{copy_num}
holds the ancestral copy number indicator which will have values like
"A0B1C1" or "A2B0C0", etc.}
}
\description{
This just sums up the length of genome from each of the ancestries.
}
