% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry_barplot.R
\name{ancestry_barplot}
\alias{ancestry_barplot}
\title{Very simple summaries of extended ancestry segments}
\usage{
ancestry_barplot(
  E,
  plot_type = "fracts",
  diagnostic_markers = NULL,
  variable_markers = NULL,
  inds_per_row = 400,
  sort_spec = c("desc(QA)"),
  admixture_line = "A",
  catOrd = c("A2B0C0", "A1B1C0", "A1B0C1", "A0B1C1", "A0B0C2", "A0B2C0"),
  fill_scale = scale_fill_brewer(type = "div", palette = "RdYlBu")
)
}
\arguments{
\item{E}{the tibble of extended ancestry segments.
Has the format of the output of extend_ancestral_segments_3()}

\item{plot_type}{string saying what type of plot to make: "fracts" just
plots the overall fractions of ancestry copy-numbers in each individual,
while "segments" actually places the ancestry copy-number segments along
the vertical axis according to where they are in the genome.}

\item{diagnostic_markers}{A tibble of species-diagnostic markers.  Must have
a column \code{diag_spp} which holds a value of A, B, or C, \code{chrom}, \code{pos}, \code{indiv},
and \code{n}.  \code{n} is the number of copies of the species diagnostic allele at the
locus.  Obviously the chromosome names and the sample names must match those
in E.}

\item{variable_markers}{a tibble that holds the variable markers. Must
have a column \code{var_spp} which holds a value of A, B, or C, naming the
species in which the marker is known to be variable.}

\item{inds_per_row}{Number of indviuals on each row of the plot}

\item{sort_spec}{string that would be an input to \code{arrange()} in order
to sort individuals in certain ways.  The default is by the admixture
proportion of the "A" species, i.e. \code{QA}.  You could use \code{QB} or \code{QC},
or, if you wanted to sort by the fractions of different full ancestry patterns,
you could use, for example:
\code{sort_spec <- c("desc(A2B0C0)", "desc(A1B1C0)", "desc(A1B0C1)", "desc(A0B2C0)", "desc(A0B1C1)", "desc(A0B0C2)")}.
But, it is probably best to just sort by admixture fractions. That way the admixture fraction
line gets plotted reasonable.}

\item{admixture_line}{string giving the Q value to plot on the line that overlaps the
columns.  Could be \code{A}, \code{B}, or \code{C}.  If it is NULL, then null line is plotted
showing the admixture fraction.}

\item{catOrd}{the order in which you want the colors of the different full
ancestry pattern categories to appear in the columns.}
}
\description{
Just a small function to make some summaries of
a tibble of extended segments across a lot of
individuals.  To be used to get a sense for how
much variation there is across individuals in
ancestry.
}
