% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_HOT}
\alias{calc_HOT}
\title{Count the number of homozygous opposites between a kid and candidate parents}
\usage{
calc_HOT(G, kid, par)
}
\arguments{
\item{G}{a integer matrix with loci in rows and samples in columns.  0,1,2 for homozygous
reference, heterozygote, and homozygous alternate.  -1 denotes missing data.}

\item{kid}{the 0-based subscript of the focal kid}

\item{par}{integer vector of the 0-based subscripts of the candidate parents you
want to compare the kid to.}
}
\value{
Retuns a list that can be easily converted to a tibble.  It has four vector
elements: kIdx (0-based index of kid) pIdx (0-based index of candidate parent), num_non_missing,
and num_hot.
}
\description{
This is included as a place to have this calculation for comparison, but is
not actually used in the MixedUpParents methodology.
}
