% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromo_segs.R
\name{chromo_segs}
\alias{chromo_segs}
\title{Collapse runs of adjacent same-numbers of ancestral alleles into segments}
\usage{
chromo_segs(D, chrom_levels)
}
\arguments{
\item{D}{a tibble with columns:
\itemize{
\item \code{chrom}: the name of the chromosome
\item \code{pos}: the position along that chromosome
\item \code{n}: the number of copies (0, 1, or 2, or NA) of the species-specific alleles
carried by this individual at this position.
}}

\item{chrom_levels}{the different chromosome values in order. (This function
will create a factor chrom_f)}
}
\value{
This returns a tibble that has the following columns:
\itemize{
\item \code{chrom}: the chromosome
\item \code{chrom_f}: the chromosome as a factor with levels from \code{chrom_levels}
\item \code{n}: the number of copies of the species-diagnostic allele
\item \code{n_f}: n as a factor rather than an integer (useful for plotting with categorical colors)
\item \code{start}: the position of the marker at which the segment begins.
\item \code{stop}: the position of the marker at which the segement ends
}
}
\description{
This is designed to operate on a single individual and with reference to
a single diagnostic species.
}
