% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diag_markers_10_fish}
\alias{diag_markers_10_fish}
\title{Marker data with of 10 fish from 3 sets of species-diagnostic markers}
\format{
A tibble like
}
\source{
Jared passed them to me
}
\description{
Marker data with of 10 fish from 3 sets of species-diagnostic markers
}
