% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_allele_freqs_EM.R
\name{estimate_allele_freqs_EM}
\alias{estimate_allele_freqs_EM}
\title{Estimate allele frequencies in the different backgrounds by EM}
\usage{
estimate_allele_freqs_EM(E, V)
}
\arguments{
\item{E}{the ancestral copy number segs tibble like that produced
by extend_ancestral_segments_3().}

\item{V}{the variable-markers genotypes. Must have indiv IDs and
chromosome names that are consistent with those in E.}
}
\description{
Takes the ancestral copy-number segs tibble and the variable
markers data and joins and filters them and then counts up
alleles in different ancestry copy number segments and then
simply uses R to implement an EM algorithm.
}
