% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{id}
\alias{id}
\title{interval difference with a shorter name and default no check validity}
\usage{
id(a, b)
}
\arguments{
\item{a}{an Interval object}

\item{b}{another Interval object}
}
\description{
interval difference with a shorter name and default no check validity
}
\keyword{internal}
