% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{ii}
\alias{ii}
\title{interval intersection with a shorter name and default no check validity}
\usage{
ii(a, b)
}
\arguments{
\item{a}{an Interval object}

\item{b}{another Interval object}
}
\description{
interval intersection with a shorter name and default no check validity
}
\keyword{internal}
