% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer_representation_EAS.R
\name{integer_representation_EAS}
\alias{integer_representation_EAS}
\title{Create an integer representation of everyone's extended ancestry segments}
\usage{
integer_representation_EAS(E)
}
\arguments{
\item{E}{The output of \code{extend_ancestral_segments_3()}. This table should
include all the individuals that you interested in doing pairwise comparisons
between.}
}
\value{
Returns a list with six components:
\itemize{
\item \code{nIndiv}: the number of individuals
\item \code{nChrom}: the number of chromosomes
\item \code{iKey}: a tibble giving the base-0 indexes of the individuals (and their names)
\item \code{cKey}: a tibble giving the base-0 indexes of the chromosomes. This and the previous
output are useful for reassociating the sample and chromosome names with later output.
\item \code{IntervalsMatrix}: an integer matrix with five columns: start, stop,
anc_int, iIdx, and cIdx.  Each row corresponds to an interval (or a chromosome with no
intervals) in and individual.  iIdx and cIdx are not really needed since we
also have the index, but I am going to leave those on there for now just
while developing. I put them on the end so that if I drop them it won't change
the subscripting of the columns at all.
-\code{IndexMatrix}: An integer matrix with columns first, last, iIdx, and cIdx.  Again,
iIdx and cIdx are not needed and I will likely drop them once I get used to
indexing arrays with these.
}

As for indexing arrays, here is how it works.  Imagine you want to get the
intervals for sample i on chrom c (where both i and c are zero-based indexes).
Let the number of chromosomes be nC and let the IndexMatrix be called \code{X}, then the
lo and hi indexes (base-0) of the rows in IntervalMatrix that hold those intervals
are given (in base-0 subscripting, for C++) by \code{X(i * nC + c, 0)} and \code{X(i * nC + c, 1)}.
By this system we can quickly go through lots of chromosomes in different individuals
and intersect the intervals thereon.
}
\description{
This integer representation will be passed to RCpp for fast intersection
of segments between pairs of individuals.  We are basically giving individuals
and chromosomes base-0 integer indexes.
}
