% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{intersect_ancestry_intervals_rcpp}
\alias{intersect_ancestry_intervals_rcpp}
\title{Intersect the ancestry intervals of individuals}
\usage{
intersect_ancestry_intervals_rcpp(grp1, grp2, nC, V1, V2, X1, X2, nv1, nv2)
}
\arguments{
\item{grp1}{The 0-based indices of the individuals in the first group of indviduals to intersect}

\item{grp2}{The 0-based indices of the individuals in the second group}

\item{nC}{The number of chromosomes}

\item{V1}{The Interval Matrix for the group 1 individuals}

\item{V2}{The Interval Matrix for the group 2 individuals}

\item{X1}{The index for V1}

\item{X2}{The index for V2}

\item{nv1}{vector of number of intervals for each individual in V1}

\item{nv2}{vector of number of intervals for each individual in V1}
}
\value{
This returns a matrix that is somewhat like the input V1 or V2.
It has columns of Idx1 (0), Idx2 (1), cIdx (2), anc1 (3), anc2 (4), start (5), stop (6).  This holds
the intersected intervals (or -1 if none exist on the chromosome).
}
\description{
This is a low level function that should probably not be directly
accessed by users.  It is designed to be relatively general and I think
it will be better to wrap it within an R function for the user, later.
}
