% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IntervalsEmpty}
\alias{IntervalsEmpty}
\title{An empty interval object saved in data for convenience}
\format{
An interval object
}
\source{
I created it
}
\description{
An empty interval object saved in data for convenience
}
