% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{iu}
\alias{iu}
\title{interval union with a shorter name and default no check validity}
\usage{
iu(a, b)
}
\arguments{
\item{a}{an Interval object}

\item{b}{another Interval object}
}
\description{
interval union with a shorter name and default no check validity
}
\keyword{internal}
