% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedUpParents-package.R
\docType{package}
\name{MixedUpParents-package}
\alias{MixedUpParents}
\alias{MixedUpParents-package}
\title{MixedUpParents: Efficient, Accurate Parentage Inference in Admixed Populations}
\description{
Uses markers with alleles diagnostic for different species to infer (via an heuristic algorithm) the number of doses (0, 1, or 2) of ancestry from each species at different segments within each individual's genome. Uses those "ancestry copy numbers" to estimate allele frequencies in each species at a set of variable/polymorphic markers. Then uses all of the above to calculate the likelihood that any pair is parental vs unrelated.
}
\author{
\strong{Maintainer}: Eric C. Anderson \email{eric.anderson@noaa.gov} (\href{https://orcid.org/0000-0003-1326-0840}{ORCID})

}
\keyword{internal}
