% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pgp_rcpp}
\alias{pgp_rcpp}
\title{Low level function to compute the pairwise genotype probabilities}
\usage{
pgp_rcpp(IXG, AF, isD, AD)
}
\arguments{
\item{IXG}{an integer matrix of the intersected intervals with columns
\itemize{
\item kIdx: 0-based index of the kid
\item pIdx: 0-based index of candidate parent
\item cIdx: 0-based index of chromosome
\item anck: ternary trit giving ancestry of kid at this segment
\item ancp: ternary trit giving ancestry of candidate parent at this segment
\item start: integer base pair of the start of the segment. Will probably be used just
to know when we are reaching a new segment.
\item stop: integer base pair of the stop of the segment (probably won't be used)
\item mIdx: 0-based index of the marker
\item pos:  base-pair position of the marker (likely will not be used)
\item gk:   count of 1 alleles in the kid genotype (0, 1, or 2, or -1 for missing data)
\item gp:   count of 1 alleles in the parent genotype (0, 1, or 2, or -1 for missing data)
}}

\item{AF}{a numeric matrix with number of columns equal to the number of ancestries and
ordered like A, B, C,... and number of rows equal to the number of variable and diagnostic
markers combined.}

\item{isD}{an integer vector of 0s and 1s with length equal to the number of
variable and diagnostic markers combined. A 1 means the marker is a species-diagnostic
marker and a 0 means otherwise.}

\item{AD}{a numeric matrix with number of columns equal to the number of ancestries
and number of rows equal to the total number of individuals that got put into the
integer representation. Each row sums to one.  These are the admixture fractions.}
}
\description{
This should never really be used directly by users.  Rather, use the
\code{pairwise_genotype_probs()} function that calls this.
}
