% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ancestral_segs.R
\name{plot_ancestral_segs}
\alias{plot_ancestral_segs}
\title{Make a plot of the ancestral segments with 0, 1, or 2 copies of a species-specific allele}
\usage{
plot_ancestral_segs(
  D,
  M,
  spp_levels = unique(D$diag_spp),
  SF = 0.66,
  MOP = 0.2,
  E = NULL
)
}
\arguments{
\item{D}{a tibble like the one that comes out of \code{ancestral_n_segments()}.}

\item{M}{a tibble of the markers like in \code{\link{diag_markers_10_fish}}}

\item{spp_levels}{a vector with the order of the species desired from top
to bottom for each chromosome. Defaults to \code{unique(D$diag_spp)}.}

\item{SF}{fraction of a chromosome's vertical space occupied by the species specific
markers and segments (as opposed to being a blank break between chromosomes)}

\item{MOP}{"marker overhang part", the fraction of the vertical space devoted to markers
and segments that is accounted for by the marker "whisker" that extends from either
the top of the bottom of the segment}

\item{E}{the output from \code{extend_ancestral_segments_3()} if you want to plot
those as well.  Must have been made from D.}
}
\description{
Make a plot of the ancestral segments with 0, 1, or 2 copies of a species-specific allele
}
\examples{
data(diag_markers_10_fish)
M <- diag_markers_10_fish
D <- ancestral_n_segments(M, unique(M$chrom))
g <- plot_ancestral_segs(D, M)

# with many individuals, it is recommended to plot this large:
# like: ggsave(g, filename = "gplot.pdf", width = 30, height = 40)

# If you want to add the extended segments to it:
# E <- extend_ancestral_segments_3(D, c("WCT", "RBT", "YCT"))
# g2 <- plot_ancestral_segs(D, M, E = E)

#' # with many individuals, it is recommended to plot this large:
# like: ggsave(g2, filename = "E-gplot.pdf", width = 30, height = 40)
}
