% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_gpp.R
\name{prepare_for_gpp}
\alias{prepare_for_gpp}
\title{Collect inputs for calculating genotype pair probabilities}
\usage{
prepare_for_gpp(V, M, E, diag_err_rate = 0.005)
}
\arguments{
\item{V}{Long tibble of variable-marker genotypes with columns
\code{chrom}, \code{pos}, \code{indiv}, \code{n}.}

\item{M}{long tibble of diagnostic markers with columns \code{diag_spp},
\code{chrom}, \code{pos}, \code{indiv}, \code{n}}

\item{E}{The tibble of ancestral segments as it comes out of
\code{extend_ancestral_segments_3()}.}

\item{diag_err_rate}{the rate at which diagnostic alleles might not be
so diagnostic for a given species.  Defaults to 1 in 200.}
}
\value{
This function retuns a list with a number of components.  This
list serves as one of the inputs to \code{pairwise_genotype_probs()}. The components
of the return list are:
\itemize{
\item \code{mKey}: a tibble showing the integer equivalent of chrom-key + "-" position
for indexing the markers.
\item \code{AF}: A tibble of allele frequencies.
\item ``
}
}
\description{
The purpose of this is to collect the necessary inputs and prepare
everything to be passed into another function that will do the
intersecting of intervals and then the calculation of genotype
pair probabilities.
}
\details{
The things done by this function are:
\itemize{
\item Estimating allele frequencies of the variable markers in the
different backgrounds, using an EM algorithm.
\item Estimating admixture fractions of the individuals using the
extended ancestral segments in each of them.
\item Creating an integer representation of the extended segments that
can be used for efficiently intersecting segments.
}

The \code{V} and \code{M} inputs are largely the "raw" long format inputs,
and \code{E} is what comes out of \code{extend_ancestral_segments_3()}.
}
