% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{tib2interval}
\alias{tib2interval}
\title{turn a tibble with start and stop columns into an Intervals object}
\usage{
tib2interval(tib)
}
\arguments{
\item{tib}{a tibble with the columns start and stop}
}
\description{
turn a tibble with start and stop columns into an Intervals object
}
\keyword{internal}
