% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trit2vec}
\alias{trit2vec}
\title{Helper function to turn trits to a vector of one or two ancestries}
\usage{
trit2vec(x)
}
\description{
As a reminder:
\itemize{
\item 2 is two copies of the 1-ancestry
\item 4 is one copy of the 2-ancestry and one of the 1-ancestry
\item 6 is two copies of the 2-ancestry
\item 12 is one copy of the 2-ancestry and one copy of the 3-ancestry
\item 10 is one copy of 3 and one copy of 1
\item 18 is two copies of the 3-ancestry
This is currently configured to deal with up to 3 ancestries (MaxAnc = 3)
But this can easily be changed by setting MaxAnc = 4 or 5, etc.
}
}
\examples{
trits <- c(2, 4, 6, 10, 12, 18)
names(trits) <- trits
lapply(trits, trit2vec)
}
