% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_genotype_probs.R
\name{pairwise_genotype_probs}
\alias{pairwise_genotype_probs}
\title{Calculate the genotype probablities for a pair of individuals}
\usage{
pairwise_genotype_probs(L, kid, par)
}
\arguments{
\item{L}{the list output from \code{prepare_for_gpp()}.}

\item{kid}{the ID of the individual (the actual ID, not the iIdx) to be treated
as the kid.}

\item{par}{a vector of the actual IDs of the individuals to be treated as parents}
}
\description{
More description later
}
